Javascript + Unity builds now sync up
- Cursors UI settings got changed to make them anchor to the Game Area the same way as the Javascript build (Top left corner is 0,0).
- Shockwaves needed to use the same anchor logic but it was a Particle System, so I took that away and made a scripted animation of the ripple.

+++ RelayTestApp/Assets/Scripts/Managers/BrainCloudManager.cs
@@ -364,14 +364,14 @@ public class BrainCloudManager : MonoBehaviour
                         if (op == "move")
                         {
                             member.IsAlive = true;
-                            member.MousePosition.x = Convert.ToSingle(data["x"]);
-                            member.MousePosition.y = -Convert.ToSingle(data["y"]); // + _mouseYOffset;
+                            member.MousePosition.x = (float)Convert.ToDouble(data["x"]);
+                            member.MousePosition.y = (float)-Convert.ToDouble(data["y"]); // + _mouseYOffset;
                         }
                         else if (op == "shockwave")
                         {
                             Vector2 position; 
-                            position.x = Convert.ToSingle(data["x"]);
-                            position.y = -Convert.ToSingle(data["y"]);
+                            position.x = (float)Convert.ToDouble(data["x"]);
+                            position.y = (float)-Convert.ToDouble(data["y"]);
                             member.ShockwavePositions.Add(position);
                         }
                     }
@@ -384,14 +384,14 @@ public class BrainCloudManager : MonoBehaviour
                         if (op == "move")
                         {
                             member.IsAlive = true;
-                            member.MousePosition.x = Convert.ToSingle(json["x"]);
-                            member.MousePosition.y = -Convert.ToSingle(json["y"]); // + _mouseYOffset;
+                            member.MousePosition.x = (float)Convert.ToDouble(json["x"]);
+                            member.MousePosition.y = (float)-Convert.ToDouble(json["y"]); // + _mouseYOffset;
                         }
                         else if (op == "shockwave")
                         {
                             Vector2 position; 
-                            position.x = Convert.ToSingle(json["x"]);
-                            position.y = -Convert.ToSingle(json["y"]);
+                            position.x = (float)Convert.ToDouble(json["x"]);
+                            position.y = (float)-Convert.ToDouble(json["y"]);
                             member.ShockwavePositions.Add(position);
                         }
                     }
